[MODULES-383] override org.jboss.modules.JDKSpecific#getPlatformClassLoader()
This override changes the behavior of JBoss Modules on JDK 9+
to make sure that the parent classloader of `ModuleClassLoader`s
is not the system classloader, but the classloader that loaded
JBoss Modules themselves.

We already override several JBoss Modules classes in the Thorntail
codebase. This override is added here because we only need to change
the JDK 9+ variant of the method, and this project is already set up
for building with JDK 9+. The main Thorntail project is intentionally
kept on Java 8. The main project just needs to make sure that this
artifact is unpacked _on top_ of already unpacked JBoss Modules.